backend/requirements.txt
Django==3.2.16
mixer==7.1.2
Pillow==8.3.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
Faker==12.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
PyJWT==2.1.0
webcolors==1.13
django-filter==23.3
djoser==2.1.0
python-dotenv==0.20.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
gunicorn==20.1.0